+++ cnapy/core.py
@@ -4,15 +4,11 @@ import itertools
 from collections import defaultdict
 from typing import Dict, Tuple, List
 from collections import Counter
-import gurobipy
-import io
-import traceback
 import numpy
 import cobra
 from cobra.util.array import create_stoichiometric_matrix
 from cobra.core.dictlist import DictList
 from optlang.symbolics import Zero, Add
-from qtpy.QtWidgets import QMessageBox
 
 import efmtool_link.efmtool4cobra as efmtool4cobra
 import efmtool_link.efmtool_extern as efmtool_extern
@@ -363,45 +359,3 @@ def replace_ids(dict_list: DictList, annotation_key: str, unambiguous_only: bool
                 pass
         if len(candidates) > 0 and old_id == entry.id:
             print("Could not find a new ID for", entry.id, "in", candidates)
-
-
-# TODO: should not be in the core module
-def except_likely_community_model_error() -> None:
-    """Shows a message in the case that using a (size-limited) community edition solver version probably caused an error."""
-    community_error_text = "Solver error. One possible reason: You set CPLEX or Gurobi as solver although you only use their\n"+\
-                            "Community edition which only work for small models. To solve this, either follow the instructions under\n"+\
-                            "'Config->Configure IBM CPLEX full version' or 'Config->Configure Gurobi full version', or use a different solver such as GLPK."
-    msgBox = QMessageBox()
-    msgBox.setWindowTitle("Error")
-    msgBox.setText(community_error_text)
-    msgBox.setIcon(QMessageBox.Warning)
-    msgBox.exec()
-
-
-def get_last_exception_string() -> str:
-    output = io.StringIO()
-    traceback.print_exc(file=output)
-    return output.getvalue()
-
-
-def has_community_error_substring(string: str) -> bool:
-    return ("Model too large for size-limited license" in string) or ("1016: Community Edition" in string)
-
-
-def model_optimization_with_exceptions(model: cobra.Model) -> None:
-    try:
-        return model.optimize()
-    except gurobipy.GurobiError as error:
-        msgBox = QMessageBox()
-        msgBox.setWindowTitle("Gurobi Error!")
-        msgBox.setText("Calculation failed due to the following Gurobi solver error " +\
-                       "(if this error cannot be resolved,\ntry using a different solver by changing " +\
-                       "it under 'Config->Configure cobrapy'):\n"+error.message+\
-                       "\nNOTE: Another error message will follow, you can safely ignore it.")
-        msgBox.setIcon(QMessageBox.Warning)
-        msgBox.show()
-    except Exception:
-        exstr = get_last_exception_string()
-        # Check for substrings of Gurobi and CPLEX community edition errors
-        if has_community_error_substring(exstr):
-            except_likely_community_model_error()
